+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # V.0.3.1
 
 
-## Bug Fixes
-- Add source file name in track file for merge command
+## Changes
+- No need GTF be sorted anymore.
+- Merge and Classify now support create index on the fly if not provided.
+
+
+